Refactor seeding and db connection
